Fixed timer_hpp

+++ rover_auxiliary/src/aruco_detection/detection.hpp
@@ -3,7 +3,7 @@
 
 #include "frame_processing.hpp"
 #include "rovus_lib/moving_average.hpp"
-//#include "rovus_lib/timer.hpp"
+#include "rovus_lib/timer.hpp"
 #include <unordered_map>
 
 #define COEFF_NB_ARUCO 10
@@ -19,7 +19,7 @@ class ArucoDetection
   private:
 
     static constexpr float VALIDATION_THRESHOLD = 0.75;
-    static constexpr uint8_t DELAY_BETWEEN_CAPTURE_MS = 100;
+    static constexpr uint64_t DELAY_BETWEEN_CAPTURE_MS = 100;
     FrameProcessing processedFrame;
     std::unordered_map<uint16_t, MovingAverage<uint16_t, COEFF_NB_ARUCO>> validation;
 };
